rename: 이미지 경로 enum 패키지 이동
